requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.0
django-celery-beat==2.7.0
django-push-notifications==3.1.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.15.2
flower==2.0.1
googlemaps==4.10.0
h11==0.14.0
httptools==0.6.4
humanize==4.12.1
idna==3.10
kombu==5.4.2
msgpack==1.1.0
pika==1.3.2
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
pystache==0.6.7
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
static==1.1.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
